更新UI， 调整一些代码

+++ client/UnityProject/Assets/Scripts/Data/Static/Sheet/LanguagesData.cs
@@ -95,7 +95,7 @@ public class LanguagesDataManager
 		}
 	}
 
-	public LanguagesData Get(UInt16 key)
+	public LanguagesData Get(Byte key)
 	{
 		if (!hashTable.ContainsKey(key))
 		{
